Fix ValueError: Buffer dtype mismatch, expected 'long' but got 'int'
Change np.int32 to np.int64 for node and edge index arrays to match
the Cython cyaccess extension expectations on 64-bit systems.

+++ pandana/network.py
@@ -92,8 +92,8 @@ class Network:
         )
 
         # Keep integer buffers aligned with C long expected by cython extension.
-        node_idx_values = np.asarray(self.node_idx.values, dtype=np.int32)
-        edge_idx_values = np.asarray(edges.values, dtype=np.int32)
+        node_idx_values = np.asarray(self.node_idx.values, dtype=np.int64)
+        edge_idx_values = np.asarray(edges.values, dtype=np.int64)
 
         self.net = cyaccess(
             node_idx_values,
